consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["WeWe-power <84448872+WeWe-power@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-binance = "^1.0.19"
pika = "^1.3.2"
pydantic = "^2.6.1"
redis = "^5.0.1"
websocket-client = "^1.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
